MNT: refactor to separate main command logic from parser creation
